Refactor endpoint constant imports to use correct casing in auth and settings services
